Extracts router functions in separate composable
